Move dialog creation to helper file
